add createweight to db store

+++ db/migrations/20190626204848_CreateWeight.sql
@@ -3,7 +3,7 @@
 -- SQL in section 'Up' is executed when this migration is applied
 
 
-CREATE TABLE weight (id SERIAL PRIMARY KEY, app_user_id INTEGER REFERENCES app_user(id), weight DECIMAL, weight_date DATE);
+CREATE TABLE weight (id SERIAL PRIMARY KEY, app_user_id INTEGER REFERENCES app_user(id), weight NUMERIC(4, 1), weight_date DATE);
 
 
 -- +goose Down
